Add what to log to the config (see description)
The order of the items this commit ads correlates to the id it will have
in the _log table in the database, under the "type" column. For
instance, when a player runs a command, it it classed as type 2.

Below is a list of what each type ID correlates to:
1 - chat
2 - command
3 - join
4 - quit
5 - kick
6 - serverchange
7 - namechange